// src/app/models/asset-exchanging-event.ts
export interface AssetExchangingEvent {
    eventId: string;
    requestOrgId: string;
    requestOrgIdentifier: string;
    requestOrgNam: string;
    org2: string;
    targetOrgId: string;
    targetOrgIdentifier: string;
    targetOrgNam: string;
    status: string;
    assetId: string;
    assetName: string;
    message: string;
    exchangeOrgId: string;
    exchangeOrgIdentifier: string;
    exchangeOrgNam: string;
    dateTimeFromNow: string;
}